requirements.txt
stramlit
langchain
langchain_openai
openai
wikipedia
chromadb
tiktoken
python-dotenv